+++ CHANGELOG.md
@@ -1,7 +1,10 @@
 # Changelog
 
-## [Dev - 0.9.0] 2020-08-29
+## [Dev - 0.9.0] 2020-09-03
+- Added sign in tutorial screen.
+- Fixed crashes when clicking the resend email link.
 
+## [Dev - 0.9.0] 2020-08-29
 - Added splash screen.
 - Added login screen.
 - Added Crashlytics.